Extract a list of steps view to a separate view `ShippingLabelPrintingStepsView`.
